edit

+++ src/main/java/com/tamerb/ecommerce/model/Product.java
@@ -10,7 +10,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer productID;
+    private Long productID;
 
     private @NotNull String name;
